This patch ports the MySQL Shell code to be built using a MariaDB server as main dependency, ensuring the available functionality is valid for both MySQL and MariaDB Servers.
Due to the implementation differences in the MariaDB ecosystem the following
functionality has been eliminated:

EXCLUDED FUNCTIONALITY
======================
- JavaScript support
- X Protocol/Dev API
- Admin API
- Dump And Load
- Upgrade Checker
- Login Path Credential Helper
- MySQL Authentication Plugins

From a MySQL Shell user perspective, the following elements will not be
available:

X Protocol/Dev API
------------------
- mysqlx module
- All the DevAPI (Session, Schema, Collection, Table, ...)

Admin API Components
--------------------
- The global objects were removed: dba, cluster, replicaset
- mysqlsh.connect_dba()
- shell.conenct_to_primary()

Dump and Load Components
------------------------
- util.importJson()
- util.importTable()
- util.loadDump()
- util.exportTable()
- util.dumpTables()
- util.dumpSchemas()
- util.dumpInstance()
- util.copyInstance()
- util.copySchemas()
- util.copyTables()
- util.dumpBinlogs()
- util.loadBinlogs()

Upgrade Checker Components
--------------------------
- util.check_for_server_upgrade()

Removed Command Line Options
----------------------------
Shell Options
-------------
--mx
--mysqlx
--sqlx
--redirect-primary
--redirect-secondary
--cluster
--replicaset
--js
--javascript
--js-debug-port=<value>
--js-debug-wait-attached

+++ unittest/scripts/auto/py_shell/scripts/credential_handling_norecord.py
@@ -9,7 +9,10 @@ shell.delete_all_credentials()
 # Create the sample user
 shell.connect(__mysqluripwd)
 session.run_sql("DROP USER IF EXISTS sample@localhost")
-session.run_sql("CREATE USER sample@localhost IDENTIFIED WITH caching_sha2_password by 'password'")
+if sandbox.vendor() == "MariaDB":
+    session.run_sql("CREATE USER sample@localhost IDENTIFIED by 'password'")
+else:
+    session.run_sql("CREATE USER sample@localhost IDENTIFIED WITH caching_sha2_password by 'password'")
 session.run_sql("GRANT ALL on *.* to sample@localhost")
 shell.disconnect()
 
@@ -51,7 +54,7 @@ EXPECT_STDOUT_CONTAINS(f"<ClassicSession:{sample_user}>")
 testutil.call_mysqlsh(parameters, "", ["MYSQLSH_TERM_COLOR_MODE=nocolor"], "mysqlshrec")
 EXPECT_STDOUT_CONTAINS(f"<ClassicSession:{sample_user}>")
 
-#@<> Dump, using the stored password just once for the initial connection, not for the dump operation
+#@<> Dump, using the stored password just once for the initial connection, not for the dump operation {__have_dump_and_load}
 shell.connect(__mysqluripwd)
 session.run_sql('drop schema if exists connection_handling')
 session.run_sql('create schema connection_handling')
@@ -61,7 +64,7 @@ WIPE_SHELL_LOG()
 testutil.call_mysqlsh(parameters[:5] + ["--", "util", "dump-instance", "testing"], "", ["MYSQLSH_TERM_COLOR_MODE=nocolor"], "mysqlshrec")
 EXPECT_SHELL_LOG_CONTAINS_COUNT("Retrieving password from credential manager", 1)
 
-#@<> Load, using the stored password just once for the initial connection, not for the load operation
+#@<> Load, using the stored password just once for the initial connection, not for the load operation {__have_dump_and_load}
 shell.connect(__mysqluripwd)
 session.run_sql('drop schema connection_handling')
 session.close()
@@ -70,7 +73,8 @@ testutil.call_mysqlsh(parameters[:5] + ["--", "util", "load-dump", "testing"], "
 EXPECT_SHELL_LOG_CONTAINS_COUNT("Retrieving password from credential manager", 1)
 
 #@<> Cleanup
-testutil.rmdir('testing', True)
+if __have_dump_and_load:
+    testutil.rmdir('testing', True)
 shell.options["credentialStore.helper"] = current_helper
 shell.connect(__mysqluripwd)
 session.run_sql("DROP USER IF EXISTS sample@localhost")
