+++ unittest/scripts/auto/py_shell/scripts/init_scripts_norecord.py
@@ -19,16 +19,17 @@ if not mysqlsh_dir_exists:
     os.mkdir(mysqlsh_dir)
 
 # Ensures a startup file for JS and PY exist
-if not init_js_exists:
+if not init_js_exists and __have_js:
     testutil.create_file(init_js, "print('Loaded JS Startup File')")
 if not init_py_exists:
     testutil.create_file(init_py, "print('Loaded PY Startup File')")
 
 # Create test files to verify both startup and mode switch on each case
-testutil.create_file('test.js', "\\py")
-testutil.create_file('test.py', "\\js")
+if __have_js:
+  testutil.create_file('test.js', "\\sql")
+testutil.create_file('test.py', "\\sql")
 
-#@<> Tests initialization and switch to python
+#@<> Tests initialization and switch to sql {__have_js}
 WIPE_OUTPUT()
 testutil.call_mysqlsh(["--interactive=full", "-f", 'test.js'], "", ["MYSQLSH_TERM_COLOR_MODE=nocolor"])
 EXPECT_STDOUT_CONTAINS(f"WARNING: Global startup scripts are no longer supported, '{init_js}' will be ignored.")
@@ -36,20 +37,22 @@ EXPECT_STDOUT_CONTAINS(f"WARNING: Global startup scripts are no longer supported
 EXPECT_STDOUT_NOT_CONTAINS('Loaded JS Startup File')
 EXPECT_STDOUT_NOT_CONTAINS('Loaded PY Startup File')
 
-#@<> Tests initialization and switch to javascript
+#@<> Tests initialization and switch to sql
 WIPE_OUTPUT()
-testutil.call_mysqlsh(["--interactive=full", "-f", 'test.js'], "", ["MYSQLSH_TERM_COLOR_MODE=nocolor"])
-EXPECT_STDOUT_CONTAINS(f"WARNING: Global startup scripts are no longer supported, '{init_js}' will be ignored.")
+testutil.call_mysqlsh(["--interactive=full", "-f", 'test.py'], "", ["MYSQLSH_TERM_COLOR_MODE=nocolor"])
+if __have_js:
+  EXPECT_STDOUT_CONTAINS(f"WARNING: Global startup scripts are no longer supported, '{init_js}' will be ignored.")
 EXPECT_STDOUT_CONTAINS(f"WARNING: Global startup scripts are no longer supported, '{init_py}' will be ignored.")
 EXPECT_STDOUT_NOT_CONTAINS('Loaded JS Startup File')
 EXPECT_STDOUT_NOT_CONTAINS('Loaded PY Startup File')
 
 #@<> Cleanup
-os.remove('test.js')
+if __have_js:
+  os.remove('test.js')
 os.remove('test.py')
 
 # Deletes the initialization files if were created by the test
-if not init_js_exists:
+if not init_js_exists and __have_js:
     os.remove(init_js)
 if not init_py_exists:
     os.remove(init_py)
